separate out Toolbox.Comm module
